sweeping queue changes
[#123984511]

I'm sorry - this should be smaller commits. C'est la vie.

* Introduce `AckTask` pattern where the logic required to ACK a message
  encapsulated inside the message itself rather than in the `Remove`
  function on the queue. This let's us keep the queue interface generic
  rather than leaking some implementation details of SQS.

* Start building out the `Foreman` sub-system which is responsible for
  turning tasks from the queue into runnable Jobs that are injected with
  their dependencies.

None of this is wired up to the webhook just yet as some changes still
need to be made in the `EventHandler` to prepare it for splitting
between the two components at either end of the queue.

Queuing Overview
================

Asynchronous work starts life as a `Plan` which is a struct describing
the work to be done. A `Plan` can convert itself into a `Task`. A `Task`
is a generic blob of work that can be given to the queue. This means
that it has a `Type` and a `Payload` which is normally some JSON
representation of the `Plan`. This `Task` can then be serialized and
stored for later execution.

When the `Task` is retrieved from the queue it is given to the `Foreman`
to be inflated into a runnable `Job` which embeds the deserialized
`Plan`.

  Producer:

    Plan --convert--> Task --enqueue--> Queue

  Consumer:

    Queue --dequeue--> Task --inflate (by Foreman)--> Job (embeds Plan)