Simplify diffscanner
- Remove private methods that became unnecessary
- Remove Hunk struct; it's not required
- Simplify regexes, don't capture things that don't need capturing

This refactoring makes it more clear how the scanner works, in that it
scans until it finds a file header, persisting that path, then scans
until it finds a hunk header, persisting its associated line number, and
then scans until it finds a context line within the hunk, at which point
it returns true. If it finds no context line it returns false.

What I find curious about this code is that it only increments the line
within the hunk when it finds a matching context line; those that do not
match are not counted, but perhaps should.