DiffScanner takes an io.Reader
This means we don't have to read the entire diff into memory when using
the CLI; we can scan it line by line.

For DiffScanJob we still read the entire response body into memory; this
is unfortunate but we should prioritize the CLI performance as it is
intended to be longer-lived than the DiffScanJob.