Remove false positives on non-yml files
Instead of passing just the content of a line to the matchers, pass the
whole scanners.Line. This allows us to use the added context of the file
extension when determining if something is YAML or not.

We pass the Line around as a pointer for performance sake; it is not
expected that any Matcher modifies the Line. Note that matcher.Multi
creates a new Line when it wants to change the content for the matchers
it contains. Without making a new Line, the violation handler would
receive content with the modifications.

[#128032605]

Signed-off-by: Christopher Brown <cbrown@pivotal.io>