FindOrCreate uses specific attributes
Some information in the raw JSON payload is updated, like the last
updated time of the repository, thus we shouldn't use the whole
Repository in the Where call. Instead, use only the attributes which
have a uniqueness constraint on them.

At some point we should actually update repositories separately to doing
work on them in the event that updates we care about occur; we should
aim to have an up-to-date version of the repository in the DB.

[#128790261]

Signed-off-by: Christopher Brown <cbrown@pivotal.io>