Restore fetch test for multiple changes
Since switching this to a real repository the multiple-changes part of
the ChangeDiscoverer tests were lost.

This restores it by creating commits on multiple refs (master and
topicA), and then gets the actual tip of each ref and its parent and
ensures that we're storing that information as the changes in the
database.

[#131112059]