Remove context.Context from ChangeDiscoverer
The context that was there previously wouldn't have actually helped; a
signal would have only been reacted to when work was already finished.

This approach shares the signals channel and passes it along to the work
method; this means work can know when a signal has been sent and stop
doing work when it's appropriate to do so, which in this case is between
fetching and scanning repositories.

This is real hard to test because of race conditions and the
asynchronous nature of process.Signal(), so there is no test. Sorry.