Rescanner scans and notifies more intelligently
Previously the rescanner would do a dirscan when the scanner rules
version changes on each repository that hasn't yet been scanned with the
current rules version.

This meant that we only ever rescanned from the point in time we'd
originally cloned the repository, as we never update the repository on
disk. We'd also end up re-notifying all the credentials that are
presently in the working directory of the repository.

This change makes it so we dig up all the ancestry scans that haven't
yet been run with the current rules version and repeat them exactly. If
there are any new credentials found, we notify only for those new
credentials.

[#131658363]