ChangeDiscoverer's Run() respects signals
Previously the work function was given the signals channel so that it
could be canceled while work was happening, but only after a Fetch was
successfully completed. However, sharing the signals channel between
work() and Run() meant that if work()'s select received the signal,
Run()'s select would not, and the process would never exit.

This change makes it so they continue to share the signals channel, but
when either one receives a signal, it calls a context.CancelFunc:

If work()'s select receives the signal it immediately calls cancel() and
returns safely, letting Run() react to the cancel. If Run() receives the
signal it calls the cancel func itself, resulting in the same behavior.

Signed-off-by: Kris Hicks <khicks@pivotal.io>