Don't scan merge commits, scan all parents
Merge commits may have a conflict in them, and that content may have
credentials in it, but we're not presently handling the special case
where a credential has as part of its conflict resolution the adding of
a credential that isn't found in either of the merge commit's parents.

This also fixes a bug where we would only scan one parent of a merge,
failing to go down the other parent's history.