Don't scan past known SHAs
Presently the only way we know about a specific SHA having been scanned
is because we store the SHA on the credential it was found in. We can
use this information to keep a list of unique SHAs that have been
scanned for a particular repository and rules version to know when to
not continue scanning ancestors when it's not necessary.

This will reduce the number of notifications we get between scans in the
same way that a similar memoization of SHAs was done to reduce the
number of notifications within a single scan did.

[#132716879]