Use single key for github webhook secrets
We have a different set of webhooks in staging/prod, therefore we need
some array of them rather than a specific set.

We also must strip single quotes in render-bosh-manifest:

We store a single-quoted string in lastpass which has a literal array inside it. This
is so we can configure the pipeline with a string representation of the
array (params must be strings).

In render-bosh-manifest we then must strip these single quotes.

[#132953819]

Signed-off-by: Kalai Wei <kawei@pivotal.io>