AllBlobsForRef is cancelable
This method can take a long time to exit due to the fact that it's
streaming across every single blob in a Git repository to a Sniffer.

Include a context.Context in the method call so we can know when we
should exit early.

Exiting early means returning a number less than zero from within the
Walk function on the tree. It would be better if we could return a
specific number to know when the callback decided to end early; it seems
libgit2 supports this with -7 (GIT_EUSER[1]) but this isn't supported in git2go.

To mimick GIT_EUSER we add a specific error that we can know to return
from AllBlobsForRef, which the caller can then check.

This also means that the reading half of the io.Pipe needs to be in a
goroutine so we can listen on the signals channel and cancel the context
given to AllBlobsForRef. Because of that, a WaitGroup is added so we
don't race to UpdateCredentialCount when AllBlobsForRef returns but
Sniff hasn't yet.

1: https://github.com/libgit2/libgit2/blob/89c332e41b12a72d89de40d63bc568c56a2c336a/include/git2/errors.h#L34