Store credential counts with branch names
This change is necessary to know which branches have how many
credentials in them.

It also necessitated scrapping AllBlobsForRef. The replacement,
BranchCredentialCounts, unfortunately pulls in the work of actually
doing the scanning for credentials. I don't yet have a good idea for how
to pull that out; if the callback were to be passed in it would have to
break the contract that we don't let git2go objects leak out of the
gitclient (the callback would need to know about git.Blob objects).

For this we also begin storing the credential counts as JSON on the
repository; this is so we can easily update a repository's entire
current credentials. Storing things in a separate table would be messy
when it came time to updating.

A new type, db.PropertyMap, is used to store JSON. Take note of the tags
on db.Repository's CredentialCounts member.

[#133686453]