Don't update repository when registering fetch
gorm likes to update associations on save, which is unnecessary when the
association has no updates.

To make it not do this you cannot use db.Set("gorm:save_associations",
false), because then it will try to save the foreign key as 0. Instead,
you have to set the foreign key yourself and rip off the association.

Maybe this will fix the 'null' credential_counts issue.