Use cmdflag.FileFlag more widely
This gives us a nice test that the file specified actually exists at the
cost of having to cast to string the value every time we want to use it.
I'm not sure how I feel about that, given the number of instances of
having to cast it.

Perhaps there is a better solution?