Fix timing dependent slack notifier test
The fake clock's IncrementBySeconds method can be flaky if you do not
have a listener present at all times. This is due to a race between
incrementing time and the listener subscribing to the time channel i.e.
`clock.Sleep`, `c.NewTimer().C()`. It is best practice to use either
`WaitForWatcherAndIncrement` or `WaitForNWatchersAndIncrement` in this
use case as the clock will ensure that a client is subscribed to the
time channel.