Skip credential counting for uncloned repositories
The credential counting code assumes that the repository exists on disk.
Therefore if the repository has not been cloned, we should not attempt
to count credentials for it.

Also add `scripts/start-docker` as a helper to launch an interactive
shell in the `pivotalcf/cred-alert-ci` docker image. This allows us to
run tests without having `libgit2` locked locally to the required
version.

[finishes #144634627]