Do not eat up all the memory in an infinite loop
This commit fixes some previously broken logic which would get stuck in
an infinite loop anytime githubClient.GetRepo() was called. It also adds
a unit test on that method to avoid similar issues in the future.

This bug was causing big spikes in the CPU and (subsequently) swap usage
in our production environment, causing the OOM Killer to kill our revok
process over and over again.

[finishes #151305001]

Signed-off-by: Rob Dimsdale-Zucker <rdimsdale@pivotal.io>